+++ scala/core/src/acsets/acsets.scala
@@ -7,39 +7,42 @@ import upickle.default._
 import monocle.Lens
 
 import scala.language.implicitConversions
-import scala.reflect.ClassTag
 
-/** In this file, we define nested ACSets.
-  *
-  * In a nested acset, each part of an acset has an associated acset. The schema
-  * for a nested acset has an acset schema associated with each object, which
-  * defines the schema for the acsets associated to the parts assigned to that
-  * object.
-  *
-  * A nested part of a nested acset is a list of (ob, id) pairs, which at each
-  * step tell you how to get to the next nested acset.
-  *
-  * Morphisms in a nested acset schema go from nested parts to nested parts. The
-  * acset schema that contains the morphism is the "highest point" that the
-  * morphism can go, i.e. the only way to get from a nested part to another
-  * nested part via a morphism is to go all the way up to where the morphism is
-  * defined, and then go back down to the target of the morphism.
-  */
-
+/** A trait marking objects in a [[Schema]] */
 trait Ob {
+  /** The subschema for the subacsets on parts of this type */
   val schema: Schema = SchEmpty
 
+  /** This object promoted to type for the domain/codomain of a morphism */
   def asDom() = Seq(PartType(Seq(this)))
-
 }
 
+/** A trait marking morphisms in a [[Schema]]
+  *
+  * Unlike in typical categories, morphisms can have multiple domains and
+  * multiple codomains. This is shorthand for a morphism for each pair of domain
+  * and codomain.
+  *
+  * Additionally, domains/codomains for morphisms are sequences of objects,
+  * representing paths through the schema, so a morphism goes between multiple
+  * levels of the schema.
+  */
 trait Hom extends Property {
+
+  /** The possible domains of this morphism */
   val doms: Seq[PartType]
+
+  /** The possible codomains of this morphism */
   val codoms: Seq[PartType]
 
   type Value = Part
 
-  // THIS IS A HORRIBLE HACK TO WORK IN A SPECIFIC CASE, FIX SOON
+  /** Any instance of [[Property]] needs a serializer/deserializer.
+    *
+    * This is a temporary hack that works in a specific case.
+    *
+    * @todo figure out nested acset serialization
+    */
   val rw = summon[ReadWriter[Int]].bimap(
     _.path(0)._2.id,
     i => Part(Seq((codoms(0).path(0), Id(i - 1))))
@@ -47,24 +50,40 @@ trait Hom extends Property {
 
 }
 
+/** A trait marking attributes in a [[Schema]]
+  *
+  * Unlike in Julia ACSets, we do not have AttrTypes. Rather, attributes have
+  * their associated scala type given by Property.Value.
+  */
 trait Attr extends Property {
+  /** Like [[Hom]], this can have multiple domains; we interpret this
+    * mathematically as a separate attribute for each domain, but this makes it
+    * easier to write down.
+    */
   val dom: Seq[PartType]
 }
 
+/** The type of a part in a nested acset is a sequence of objects,
+  * going down through the nested schemas.
+  *
+  * So `PartType(Seq())` is the type of the root acset part, and
+  * `PartType(Seq(Box,IPort))` refers to the type of input ports on boxes.
+  */
 case class PartType(path: Seq[Ob]) extends EntityType {
-  def extend(x: Ob) = PartType(path :+ x)
+  /** Add an object to the end of the PartType */
+  def extend(x: Ob): PartType = PartType(path :+ x)
 
-  def head = PartType(path.slice(0, 1))
-  def tail = PartType(path.tail)
+  /** The first object in the PartType */
+  def head: Ob = path(0)
 
-  def headOb = path.headOption
-  def lastOb = path.lastOption
+  /** The PartType relative to the first object */
+  def tail: PartType = PartType(path.tail)
 
+  /** Returns if `this` an extension of `that`? */
   def <(that: PartType): Boolean = that.path match
     case Seq() => true
     case Seq(thathead, thattail @ _*) =>
       head == thathead && tail < PartType(thattail)
-
 }
 
 object PartType {
@@ -72,12 +91,26 @@ object PartType {
   given seqIsPartType: Conversion[Seq[Ob], PartType] = PartType(_)
 }
 
-// The empty list refers to the acset itself
+/** A part is a path through a nested acset. If you visualize a nested
+  * acset as a tree, where each node is an acset and its children are all of
+  * its subacsets, then a part tells at each level which subacset to choose.
+  *
+  * The empty path refers to the root acset.
+  *
+  * @todo
+  *   We should have "relative" parts and "absolute" parts, just like in a
+  *   filesystem there are relative and absolute paths. I think that currently
+  *   `Part` is used for both notions, which is very confusing.
+  */
 case class Part(path: Seq[(Ob, Id)]) extends Entity {
+
+  /** All of the objects of the path */
   override val ty: PartType = PartType(path.map(_._1))
 
+  /** Provide directions to go one acset deeper */
   def extend(x: Ob, i: Id) = Part(path :+ (x, i))
 
+  /** Honestly not sure if this is a good method? */
   override def extend(e: Entity) = e match {
     case (p: Part) => Part(path ++ p.path)
     case _         => SubEntity(this, e)
@@ -86,19 +119,30 @@ case class Part(path: Seq[(Ob, Id)]) extends Entity {
   def head: Part = Part(path.slice(0, 1))
   def tail: Part = Part(path.tail)
 
+  /** Checks if `this` is more specific than `that` */
   def <(that: Part): Boolean = that.path match
     case Seq() => true
     case Seq(thathead, thattail @ _*) =>
       head == thathead && tail < Part(thattail)
 
+  /** An alias for `<` */
   def in(ptype: PartType) = ty < ptype
 }
 
+/** The schema for a nested acset.
+  *
+  * The nested part comes in because anything that implements `Ob` has another
+  * schema attached to it.
+  *
+  * @todo
+  *   What does this correspond to categorically?
+  */
 trait Schema {
   val obs: Seq[Ob]
   val homs: Seq[Hom]
   val attrs: Seq[Attr]
 
+  /** Returns the subschema found by following the path in `ty`. */
   def subschema(ty: PartType): Schema = ty.path match {
     case Nil => this
     case ob :: rest => {
@@ -107,8 +151,9 @@ trait Schema {
     }
   }
 
-  // Returns all of the homs that go into the given part type
-  // Each hom is prefixed by a path of objects needed to get to that hom
+  /** Returns all of the homs that go into the given part type
+    * Each hom is prefixed by a path of objects needed to get to that hom
+    */
   def homsInto(ty: PartType): Seq[(Seq[Ob], Hom)] = ty.path match {
     case Nil => Seq()
     case ob :: rest =>
@@ -120,39 +165,33 @@ trait Schema {
 
 }
 
-// case class CaseObject(name:String) extends Ob derives ReadWriter
-// object CaseObject {
-//   implicit def objectIsPartType(ob:CaseObject): PartType = PartType(Seq(ob))
-//   implicit def obSeqIsPartType(obs:Seq[CaseObject]): PartType = PartType(obs)
-// }
-
-// case class CaseHom(name:String,doms:Seq[CaseObject],codoms:Seq[CaseObject]) extends Hom derives ReadWriter
-
-// import scala.quoted.*
-// case class CaseAttr[T:ReadWriter](name:String,doms:Seq[CaseObject],tt:Type[T]) extends Attr with PValue[T]
-
-// implicit def
-
-// }
-// object CaseAttr {
-//   implicit attrRW: ReadWriter[CaseAttr] =
-// }
-
-// case class CaseSchema(obs: Set[CaseObject])
-
-object Schema {
-  // implicit val rw: ReadWriter[Schema] = readwriter[(Seq[Ob],Seq[Hom],Seq[Attr])].bimap(
-
-  // )
-}
-
+/** An opaque wrapper around an integer */
 case class Id(id: Int)
 
+/** Storage class for the parts corresponding to an `Ob` in a schema.
+  *
+  * This is immutable; methods that logically mutate simply return new objects.
+  *
+  * @param nextId
+  *   The id to assign to the next part added
+  *
+  * @param ids
+  *   The ids of all the parts added so far. This is a `Seq` because we care
+  *   about the ordering; when the ACSet is displayed this ordering is used when
+  *   two sprites overlap.
+  *
+  * @param acsets
+  *   The subacset corresponding to each id
+  */
 case class Parts(
     nextId: Int,
     ids: Seq[Id],
     acsets: Map[Id, ACSet]
 ) {
+  /** Add multiple new parts, with subacsets given in `partacsets`.
+    *
+    * Returns a sequence of the ids of the added parts.
+    */
   def addParts(partacsets: Seq[ACSet]): (Parts, Seq[Id]) = {
     val newIds = nextId.to(nextId + partacsets.length - 1).map(Id.apply)
     val newPd = Parts(
@@ -163,17 +202,20 @@ case class Parts(
     (newPd, newIds)
   }
 
+  /** Adds a single part with subacset `acs`, returns its id. */
   def addPart(acset: ACSet): (Parts, Id) = {
     val (p, newIds) = addParts(Seq(acset))
     (p, newIds(0))
   }
 
-  def setAcset(i: Id, acs: ACSet) = {
+  /** Set the subacset corresponding to `i` to `acs` */
+  def setAcset(i: Id, acs: ACSet): Parts = {
     this.copy(
       acsets = acsets + (i -> acs)
     )
   }
 
+  /** Remove the part with id `i` */
   def remPart(i: Id) = {
     this.copy(
       ids = ids.filterNot(_ == i),
@@ -181,6 +223,11 @@ case class Parts(
     )
   }
 
+  /** Move the id `i` to the front of the list of ids.
+    *
+    * This is used, for instance, when dragging a sprite so that the sprite goes
+    * over the other parts.
+    */
   def moveFront(i: Id) = {
     this.copy(
       ids = ids.filterNot(_ == i) :+ i
@@ -188,22 +235,41 @@ case class Parts(
   }
 }
 
+/** The part corresponding to the top-level acset itself. */
 val ROOT = Part(Seq())
 
+/** The empty schema. */
 case object SchEmpty extends Schema {
   val obs = Seq()
   val homs = Seq()
   val attrs = Seq()
 }
 
+/** A nested acset.
+  *
+  * @param schema
+  *   the schema that the acset conforms to
+  *
+  * @param props
+  *   the top-level properties. The values of morphisms, attributes, and generic
+  *   [[Property]]s are stored here.  We don't need what in Catlab we call
+  *   "subparts"; it's folded into this. For instance, if this is the subacset
+  *   for an edge, then the source and target are stored here.
+  *
+  * @param partsMap
+  *   the `Parts` object for each `Ob` in the schema. This is where the
+  *   subacsets are stored.
+  */
 case class ACSet(
     schema: Schema,
     props: PropMap,
     partsMap: Map[Ob, Parts]
 ) {
 
+  /** Get the subacset corresponding to a nested part; error if invalid */
   def subacset(p: Part): ACSet = trySubacset(p).get
 
+  /** Get the subacset corresponding to a nested part; return None if invalid */
   def trySubacset(p: Part): Option[ACSet] = p.path match {
     case Nil => Some(this)
     case (x, i) :: rest =>
@@ -212,6 +278,7 @@ case class ACSet(
         .flatMap(_.acsets.get(i).flatMap(_.trySubacset(Part(rest))))
   }
 
+  /** Check if a nested part exists in the ACSet */
   def hasPart(p: Part): Boolean = p.path match {
     case Nil => true
     case (x, i) :: rest =>
@@ -222,6 +289,7 @@ case class ACSet(
       } yield res).getOrElse(false)
   }
 
+  /** Set the subacset for a nested part */
   def setSubacset(p: Part, acs: ACSet): ACSet = p.path match {
     case Nil => {
       acs
@@ -235,25 +303,35 @@ case class ACSet(
     }
   }
 
+  /** Return all of the parts of the subacset at `i` with type `x`, along with
+    * their corresponding subacsets.
+    */
   def parts(i: Part, x: Ob): Seq[(Part, ACSet)] = {
     val ps = subacset(i).partsMap(x)
     ps.ids.map(id => (i.extend(x, id), ps.acsets(id)))
   }
 
+  /** Return all of the parts of the subacset at `i` with type `x`, without
+    * subacsets */
   def partsOnly(i: Part, x: Ob): Seq[Part] = {
     val ps = subacset(i).partsMap(x)
     ps.ids.map(id => i.extend(x, id))
   }
 
+  /** Get the value of `f` at the part `i`; errors if unset. */
   def subpart(f: Property, i: Part): f.Value = subacset(i).props(f)
 
+  /** Get the value of `f` at the part `i`; returns `None` if unset. */
   def trySubpart(f: Property, i: Part): Option[f.Value] =
     trySubacset(i).flatMap(_.props.get(f))
 
+  /** Check if the part `i` has property `f` */
   def hasSubpart(f: Property, i: Part) = trySubpart(f, i) match
     case Some(j) => true
     case None    => false
 
+  /** Adds a part of type `x` to the subacset at `p` with initial subacset
+    * `init` */
   def addPart(p: Part, x: Ob, init: ACSet): (ACSet, Part) = {
     val sub = subacset(p)
     val subschema = schema.subschema(p.ty.asInstanceOf[PartType].extend(x))
@@ -264,6 +342,8 @@ case class ACSet(
     (setSubacset(p, newSub), p.extend(x, i))
   }
 
+  /** Add several parts of type `x` to the subacset at `p` with initial
+    * subacsets given by inits. */
   def addParts(p: Part, x: Ob, inits: Seq[ACSet]): (ACSet, Seq[Part]) = {
     val sub = subacset(p)
     val subschema = schema.subschema(p.ty.asInstanceOf[PartType].extend(x))
@@ -274,24 +354,31 @@ case class ACSet(
     (setSubacset(p, newSub), ids.map(i => p.extend(x, i)))
   }
 
+  /** Convenience overload of [[addParts]] */
   def addPartsProps(p: Part, x: Ob, props: Seq[PropMap]): (ACSet, Seq[Part]) = {
     val subschema = schema.subschema(p.ty.asInstanceOf[PartType].extend(x))
     addParts(p, x, props.map(p => ACSet(subschema, p)))
   }
 
+  /** Convenience overload of [[addPart]] */
   def addPart(p: Part, x: Ob, props: PropMap): (ACSet, Part) = {
     val subschema = schema.subschema(p.ty.asInstanceOf[PartType].extend(x))
     addPart(p, x, ACSet(subschema, props))
   }
 
+  /** Convenience overload of [[addPart]] */
   def addPart(x: Ob, props: PropMap): (ACSet, Part) = addPart(ROOT, x, props)
 
+  /** Convenience overload of [[addPart]] */
   def addPart(x: Ob, init: ACSet): (ACSet, Part) = addPart(ROOT, x, init)
 
+  /** Convenience overload of [[addPart]] */
   def addPart(p: Part, x: Ob): (ACSet, Part) = addPart(p, x, PropMap())
 
+  /** Convenience overload of [[addPart]] */
   def addPart(x: Ob): (ACSet, Part) = addPart(ROOT, x, PropMap())
 
+  /** Move the part `p` to the front of its parent `Parts`. See [[Parts.moveFront]]. */
   def moveFront(p: Part): ACSet = {
     assert(p.path.length > 0)
     val (prefix, (x, i)) = (Part(p.path.dropRight(1)), p.path.last)
@@ -302,6 +389,7 @@ case class ACSet(
     setSubacset(prefix, newsub)
   }
 
+  /** Set the property `f` of part `p` to `v` */
   def setSubpart(p: Part, f: Property, v: f.Value): ACSet = {
     val sub = subacset(p)
     val newSub = sub.copy(
@@ -310,6 +398,7 @@ case class ACSet(
     setSubacset(p, newSub)
   }
 
+  /** Unset the property `f` of `p` */
   def remSubpart(p: Part, f: Property): ACSet = {
     val sub = subacset(p)
     val newSub = sub.copy(
@@ -318,6 +407,7 @@ case class ACSet(
     setSubacset(p, newSub)
   }
 
+  /** Return sequence of the parts that have property `f` set to `p` */
   def incident(p: Part, f: Hom): Seq[Part] = {
     val codom = f.codoms
       .find(c => p.ty.path.drop(p.ty.path.length - c.path.length) == c.path)
@@ -341,6 +431,7 @@ case class ACSet(
     f.doms.flatMap(dom => helper(subacset(prefix), prefix, dom.path))
   }
 
+  /** Remove a part, but not any of the other parts that might refer to it. */
   def remPartOnly(p: Part): ACSet = {
     if hasPart(p) then {
       assert(p.path.length > 0)
@@ -355,6 +446,7 @@ case class ACSet(
     }
   }
 
+  /** Remove a part and all of the other parts that refer to it. */
   def remPart(p: Part): ACSet = {
     val visited = mutable.Set[Part]()
     val queue = mutable.Queue[Part](p)
@@ -376,52 +468,72 @@ case class ACSet(
     toRemove.foldLeft(this)(_.remPartOnly(_))
   }
 
+  /** Remove all of the parts in `ps` */
   def remParts(ps: Seq[Part]): ACSet = ps match
     case Seq()             => this
     case Seq(p, rest @ _*) => this.remPart(p).remParts(rest)
 
+  /** Add the properties in `newProps` to the top-level properties. */
   def addProps(newProps: PropMap): ACSet = {
     this.copy(props = props ++ newProps)
   }
 }
 
+/** This object contains the constructor method for ACSets and also a collection
+  * of wrappers around ACSet methods in the `State` monad that allow for a
+  * quasi-imperative API for modifying ACSets purely.
+  */
 object ACSet {
+  /** Construct a new ACSet with schema `s` */
   def apply(s: Schema): ACSet = ACSet(s, PropMap())
 
+  /** Construct a new ACSet with schema `s` and top-level parts `props` */
   def apply(s: Schema, props: PropMap): ACSet =
     new ACSet(s, props, s.obs.map(ob => ob -> Parts(0, Seq(), Map())).toMap)
 
+  /** `State` wrapper around ACSet.addParts */
   def addParts(p: Part, x: Ob, props: Seq[PropMap]): State[ACSet, Seq[Part]] =
     State(_.addPartsProps(p, x, props))
 
+  /** `State` wrapper around ACSet.addPart */
   def addPart(p: Part, x: Ob, props: PropMap): State[ACSet, Part] =
     State(_.addPart(p, x, props))
 
+  /** `State` wrapper around ACSet.addPart */
   def addPart(p: Part, x: Ob): State[ACSet, Part] =
     State(_.addPart(p, x))
 
+  /** `State` wrapper around ACSet.addPart */
   def addPart(x: Ob, props: PropMap): State[ACSet, Part] = State(
     _.addPart(x, props)
   )
 
+  /** `State` wrapper around ACSet.addPart */
   def addPart(x: Ob, init: ACSet): State[ACSet, Part] = State(
     _.addPart(x, init)
   )
 
+  /** `State` wrapper around ACSet.addPart */
   def addPart(x: Ob): State[ACSet, Part] = State(_.addPart(x, PropMap()))
 
+  /** `State` wrapper around ACSet.setSubpart */
   def setSubpart(p: Part, f: Property, v: f.Value): State[ACSet, Unit] =
     State.modify(_.setSubpart(p, f, v))
 
+  /** `State` wrapper around ACSet.remSubpart */
   def remSubpart(p: Part, f: Property): State[ACSet, Unit] =
     State.modify(_.remSubpart(p, f))
 
+  /** `State` wrapper around ACSet.remPart */
   def remPart(p: Part): State[ACSet, Unit] = State.modify(_.remPart(p))
 
+  /** `State` wrapper around ACSet.remParts */
   def remParts(ps: Seq[Part]): State[ACSet, Unit] = State.modify(_.remParts(ps))
 
+  /** `State` wrapper around ACSet.moveFront */
   def moveFront(p: Part): State[ACSet, Unit] = State.modify(_.moveFront(p))
 
+  /** Returns a lens into the value of the property `f` for part `x` */
   def subpartLens(f: Property, x: Part) =
     Lens[ACSet, f.Value](_.subpart(f, x))(y => s => s.setSubpart(x, f, y))
