--- scala/core/src/ui/Form.scala
@@ -1,1 +1,0 @@
-package semagrams.ui
