+++ scala/core/src/util/TextBox.scala
@@ -5,7 +5,7 @@ import semagrams.util._
 
 import scala.collection.mutable
 
-val measurementStyle = """
+private val measurementStyle = """
 position: absolute;
 visibility: hidden;
 height: auto;
@@ -16,7 +16,14 @@ font-family: Alegreya Sans, sans-serif;
 
 private val sizeCache = mutable.Map[(String, Double), Complex]()
 
-def boxSize(text: String, fontSize: Double) = {
+/** Returns the dimensions of the bounding box for `text` at `fontSize`
+  *
+  * This caches computations, so that repeated queries are fast.
+  *
+  * @todo
+  *   we hardcode most of the style for this; we should make this more flexible.
+  */
+def boxSize(text: String, fontSize: Double): Complex = {
   if (sizeCache contains (text, fontSize)) {
     sizeCache((text, fontSize))
   } else {
