+++ scala/core/src/util/Transform.scala
@@ -1,23 +1,48 @@
 package semagrams.util
 
+/** Represents an affine transformation of the 2d plane
+  *
+  * @param z
+  *   the scaling/twisting factor
+  *
+  * @param w
+  *   the translation factor
+  *
+  * Note: scaling is applied before translation.
+  */
 case class Transform(z: Complex, w: Complex) {
-  def toSvg =
+
+  /** Returns a string representing this transform in SVG style */
+  def toSvg: String =
     s"matrix(${z.x} ${z.y} ${-z.y} ${z.x} ${w.x} ${w.y})"
 
-  def zoomBy(factor: Double) = this.copy(z = z * factor)
+  /** Returns a new transform that is zoomed in by `factor`.
+    *
+    * The old and new transforms send the origin to the same point.
+    */
+  def zoomBy(factor: Double): Transform = this.copy(z = z * factor)
 
-  def zoomAtPos(factor: Double, pos: Complex) = {
+  /** Returns a new transform that is zoomed in on `pos` by `factor`
+    *
+    * The old and new transforms send `pos` to the same point.
+    */
+  def zoomAtPos(factor: Double, pos: Complex): Transform = {
     val zp = z * factor
     Transform(zp, w + (z.x - zp.x) * pos)
   }
 
-  def apply(p: Complex) = p * z + w
+  /** Apply the transform to a point */
+  def apply(p: Complex): Complex = p * z + w
 
-  def inverse(p: Complex) = (p - w) / z
+  /** Apply inverse of the transform */
+  def inverse(p: Complex): Complex = (p - w) / z
 
-  def shift(by: Complex) = this.copy(w = w + by)
+  /** Returns a new transform that is shifted by `by` */
+  def shift(by: Complex): Transform = this.copy(w = w + by)
 }
 
 object Transform {
+
+  /** The identity affine transformation */
   val identity = Transform(1, 0)
 }
